Fix: pre-release hardening audit — endian, sign-compare, and type safety
Audit 1 (Endianness): Fixed 8 bare dereferences to arena memory that bypassed
endianness conversion macros. Changed filestuff.c etc_path to use SET_OS9L
macros for all 32-bit writes (lines 1636-1642). Updated modstuff.c adapt_le0
and buffer copy (line 752) to use SET_OS9L and memcpy for consistency.

Audit 2 (Compiler Warnings): Fixed sign-compare warnings with extra flags
(-Wformat=2 -Wsign-compare -Wpointer-arith). Changed int to uint32_t/ulong/
size_t in 6 files where unsigned comparison was intended. Reduced warnings
from 42 to 0 on audited code.

Audits 3-6: Verified sizeof, struct layout, truncation, and sign-extension
audits. Found no arena-resident structs with problematic long/pointer fields,
no unsafe time_t/off_t truncations, and no sign-extension bugs.

Build: clean, executable verified at 2.0M bytes.

+++ Source/OS9exec_core/fileaccess.c
@@ -735,7 +735,7 @@ os9err pHvolnam( _pid_, syspath_typ* spP, char* volname )
       volname[ 1 ]= NUL;     
     
     #elif defined UNIX
-      int  ii; // get the current top path as name
+      size_t ii; // get the current top path as name
       for (ii= 0; ii<strlen( spP->fullName ); ii++) {
         volname[ ii ]= spP->fullName[ ii+1 ];
         if ( ii>0 && volname[ ii ]=='/' ) { volname[ ii ]= NUL; break; }
